# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "report-agent",
    "description": "기획 문서 생성 에이전트 웹 서비스 (M2 백엔드)",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/interview/sessions": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_projects__project_id__interview_sessions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}": {
      "get": {
        "tags": [
          "interview"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_interview_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}/messages": {
      "get": {
        "tags": [
          "interview"
        ],
        "summary": "List Messages",
        "description": "트랜스크립트 + SSE 재접속 리플레이 커서 (D4/D6).",
        "operationId": "list_messages_api_interview_sessions__session_id__messages_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageOut"
                  },
                  "title": "Response List Messages Api Interview Sessions  Session Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}/kick": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Kick",
        "description": "인터뷰 시작 (FR-2.2 가설 선제시부터).",
        "operationId": "kick_api_interview_sessions__session_id__kick_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}/turn": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Turn",
        "description": "자유 텍스트 턴 — 가설 확인·반박, 뼈대 정보, 게이트 화면 밖의 잡담.",
        "operationId": "turn_api_interview_sessions__session_id__turn_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}/answers": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Answers",
        "description": "라운드 답변 제출 (FR-2.3) — 선택지 카드 클릭이 이 엔드포인트를 호출한다.",
        "operationId": "answers_api_interview_sessions__session_id__answers_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswersCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}/facts/confirm": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Facts Confirm",
        "description": "팩트 확인 게이트 (FR-2.4) — 승인 시에만 팩트 저장소에 적립된다.",
        "operationId": "facts_confirm_api_interview_sessions__session_id__facts_confirm_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactsConfirmCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/sessions/{session_id}/key-messages": {
      "post": {
        "tags": [
          "interview"
        ],
        "summary": "Key Messages",
        "description": "핵심 메시지 승인 게이트 (FR-2.6).",
        "operationId": "key_messages_api_interview_sessions__session_id__key_messages_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyMessagesCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerItem": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "option": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Option"
          },
          "free_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text"
          }
        },
        "type": "object",
        "required": [
          "index"
        ],
        "title": "AnswerItem"
      },
      "AnswersCreate": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerItem"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "AnswersCreate"
      },
...
      "FactsConfirmCreate": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "edits": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edits"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "FactsConfirmCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyMessagesCreate": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "KeyMessagesCreate"
      },
      "MessageOut": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "created_at": {
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "role",
          "kind",
          "content",
          "payload",
          "created_at"
        ],
        "title": "MessageOut"
      },
...
      "SessionCreate": {
        "properties": {},
        "type": "object",
        "title": "SessionCreate"
      },
      "SessionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "round_no": {
            "type": "integer",
            "title": "Round No"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pending_questions": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Questions"
          },
          "pending_facts": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Facts"
          },
          "pending_key_messages": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Key Messages"
          },
          "checklist": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checklist"
          },
          "key_messages_approved": {
            "type": "boolean",
            "title": "Key Messages Approved"
          },
          "hypothesis": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "phase",
          "round_no",
          "status",
          "pending_questions",
          "pending_facts",
          "pending_key_messages",
          "checklist",
          "key_messages_approved",
          "hypothesis",
          "error",
          "created_at"
        ],
        "title": "SessionOut"
      },
...
      "TurnCreate": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "TurnCreate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}